Fixed... a lot of things. Broke... a lot of things.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot/Robot.java
@@ -43,8 +43,8 @@ public class Robot {
     double lastLeftPos;
     double lastRightPos;
     double lastCenterPos;
-    public double actualX;
-    public double actualY;
+    public float actualX;
+    public float actualY;
     double deltaLeft;
     double deltaRight;
     double deltaCenter;
